requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
chromadb
huggingface-hub
sentence-transformers
transformers
torch
chainlit
python-dotenv
pypdf
tiktoken